Switched from Time to Evaluations, prettified reading/writing of results

+++ HyperOp.Experiments/Program.cs
@@ -6,12 +6,28 @@ using HyperOp.Algorithms.HyperParameterOptimization;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-//Clean-Up
-Directory.Delete("./results", true);
+
+int evaluations = 10000;
+int repetitions = 3;
+
+//Clean-Up for Debugging
+#if DEBUG
+if (Directory.Exists("./results")) { Directory.Delete("./results", true); }
+#endif
 
 if (!Directory.Exists("./results")) { Directory.CreateDirectory("./results"); }
 if (!File.Exists("./results/all_results.json")) { await ExecuteExperiments(); }
-AnalyzeResults();
+
+
+//AnalyzeResults();
+//void AnalyzeResults()
+//{
+//    if (File.Exists("./results/all_results.json"))
+//    {
+//        var content = File.ReadAllText("./results/all_results.json");
+//        var allResults = JsonSerializer.Deserialize<List<ResultDTO>>(content, new JsonSerializerOptions() { NumberHandling = JsonNumberHandling.AllowNamedFloatingPointLiterals });
+//    }
+//}
 
 async Task ExecuteExperiments()
 {
@@ -22,80 +38,46 @@ async Task ExecuteExperiments()
         new IRaceSearch(),
         new RandomSearch()
     };
-    var serializerOptions = new JsonSerializerOptions() {
-        NumberHandling = JsonNumberHandling.AllowNamedFloatingPointLiterals,
-        WriteIndented = true
-    };
+
     List<FeynmanDescriptor> feynmanDescriptors = FeynmanInstanceProvider.LoadAvailableInstances().Take(10).ToList();
-    var budget = TimeSpan.FromSeconds(30);
     var allResults = new List<ResultDTO>();
 
     foreach (var algorithm in hyperOpAlgs)
     {
         foreach (var feynmanInstance in feynmanDescriptors)
         {
-            Console.WriteLine($"Executing {algorithm.GetType().Name} on instance {feynmanInstance.GetType().Name} with budget {budget.TotalMinutes} minutes.");
-            try
+            foreach (int repetition in Enumerable.Range(1, repetitions))
             {
-                var algRes = await algorithm.Execute(feynmanInstance, new CancellationTokenSource(budget).Token);
-                allResults.Add(new ResultDTO {
-                    AlgorithmName = algorithm.GetType().Name,
-                    FeynmanInstanceName = feynmanInstance.GetType().Name,
-                    ResultDetails = algRes.Select(r => new ResultDTO.ResultDetail {
-                        Parameter = r.Item1,
-                        Population = r.Item2.Select(ind => new ResultDTO.ResultDetail.SimplifiedIndividual {
-                            Objective = ind.ObjectiveVector[0],
-                            Depth = ind.Candidate.Depth,
-                            Complexity = ind.Candidate.Complexity,
-                            Length = ind.Candidate.Length,
-                            InfixRepresentation = ind.Candidate.ToInfixString()
-                        }).ToList()
-                    }).ToList()
-                });
-                //Immediately save all available results to a file, so that we can analyze them later or resume the experiment if it was interrupted.
-                File.WriteAllText($"./results/results_{algorithm.GetType().Name}_{feynmanInstance.GetType().Name}.json", JsonSerializer.Serialize(allResults.Last(), serializerOptions));
+                Console.WriteLine($"Executing repetition {repetition} of {algorithm.GetType().Name} on instance {feynmanInstance.GetType().Name} with budget of {evaluations} evaluations.");
+                try
+                {
+                    string resultFilePath = $"./results/results_{algorithm.GetType().Name}_{feynmanInstance.GetType().Name}_{repetition}.json";
+                    if (File.Exists(resultFilePath))
+                    {
+                        Console.WriteLine($"Result file {resultFilePath} already exists. Skipping execution for this repetition.");
+
+                        //Load already existing results so we won't lose them. This is important if we want to resume the experiment after an interruption.
+                        allResults.Add(ResultHandler.ReadResultFromFile(resultFilePath));
+                        continue;
+                    }
+
+                    int seed = 42 + 100 * repetition;
+                    List<(AlgorithmParameter, Population<ExpressionTree>)> algRes = await algorithm.Execute(feynmanInstance, seed, 10000);
+                    allResults.Add(ResultHandler.ConvertToResultDTO(algRes, algorithm.GetType().Name, feynmanInstance.GetType().Name, repetition, seed));
+
+                    //Immediately save all available results to a file, so that we can analyze them later or resume the experiment if it was interrupted.
+                    ResultHandler.WriteResultToFile(resultFilePath, allResults.Last());
+                }
+                catch (NotImplementedException) { /* Simply ignore the NotImplemented for now */}
+                catch (Exception ex)
+                {
+                    // Log the error and continue with the next algorithm-instance pair, we don't want one potential failure to stop the entire experiment.
+                    Console.WriteLine($"Error executing algorithm {algorithm.GetType().Name} on instance {feynmanInstance.GetType().Name}: {ex.Message}");
+                }
             }
-            catch (NotImplementedException) { /* Simply ignore the NotImplemented for now */}
-            catch (Exception ex)
-            {
-                // Log the error and continue with the next algorithm-instance pair, we don't want one potential failure to stop the entire experiment.
-                Console.WriteLine($"Error executing algorithm {algorithm.GetType().Name} on instance {feynmanInstance.GetType().Name}: {ex.Message}");
-            }
+
         }
-
-    }
-    File.WriteAllText($"./results/all_results.json", JsonSerializer.Serialize(allResults, serializerOptions));
-}
-
-void AnalyzeResults()
-{
-    if (File.Exists("./results/all_results.json"))
-    {
-        var content = File.ReadAllText("./results/all_results.json");
-        var allResults = JsonSerializer.Deserialize<List<ResultDTO>>(content, new JsonSerializerOptions() { NumberHandling = JsonNumberHandling.AllowNamedFloatingPointLiterals });
+        ResultHandler.WriteResultsToFile($"./results/all_results.json", allResults);
     }
 }
 
-public class ResultDTO
-{
-    public string AlgorithmName { get; set; }
-    public string FeynmanInstanceName { get; set; }
-    public List<ResultDetail> ResultDetails { get; set; }
-
-    public class ResultDetail
-    {
-        public AlgorithmParameter Parameter { get; set; }
-        public List<SimplifiedIndividual> Population { get; set; }
-
-
-        public class SimplifiedIndividual
-        {
-            public double Objective { get; set; }
-            public double Depth { get; set; }
-            public double Complexity { get; set; }
-            public double Length { get; set; }
-            public string InfixRepresentation { get; set; }
-        }
-    }
-
-}
